pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bslfs"
version = "0.1.0"
description = "BSL/FS calibration toolkit"
readme = "README.md"
authors = [{ name = "Lab Team" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "typer>=0.9",
]

[project.optional-dependencies]
plot = ["matplotlib>=3.6"]
dev = [
    "pytest>=7.4",
    "black>=23.11",
    "ruff>=0.1.0",
    "pre-commit>=3.5",
]
terps = [
    "pyserial>=3.5",
]

[project.scripts]
bslfs = "bslfs.cli:run"
terps-host = "bslfs.terps.runner:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.black]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
